move ag-ui server to experimental and revert server.py to upstream main
